test: publish mollusk harness helper
