+++ Cargo.toml
@@ -188,7 +188,7 @@ mollusk-svm-programs-token = "0.13.1"
 # solana programs
 solana-axelar-operators = { version = "1.1.1", path = "programs/solana-axelar-operators", default-features = false }
 solana-axelar-gas-service = { version = "1.1.1", path = "programs/solana-axelar-gas-service", default-features = false }
-solana-axelar-memo = { path = "programs/solana-axelar-memo", default-features = false }
+solana-axelar-memo = { version = "1.1.1", path = "programs/solana-axelar-memo", default-features = false }
 solana-axelar-gateway = { version = "1.1.1", path = "programs/solana-axelar-gateway", default-features = false }
 solana-axelar-governance = { version = "1.1.1", path = "programs/solana-axelar-governance", default-features = false }
 solana-axelar-its = { version = "1.1.1", path = "programs/solana-axelar-its", default-features = false }
@@ -197,9 +197,8 @@ solana-axelar-its = { version = "1.1.1", path = "programs/solana-axelar-its", de
 solana-axelar-std = { version = "1.1.1", path = "crates/solana-axelar-std" }
 governance-gmp = { version = "0.1.4", path = "crates/governance-gmp" }
 
-# helper crates (not published - dev-dependencies only)
-mollusk-harness = { path = "helpers/mollusk-harness" }
-mollusk-test-utils = { path = "helpers/mollusk-test-utils", default-features = false }
+# helper crates
+solana-axelar-mollusk-harness = { path = "helpers/solana-axelar-mollusk-harness" }
 
 [workspace.metadata.cargo-machete]
 ignored = ["num-traits"]
