+++ deny.toml
@@ -157,12 +157,7 @@ expression = "MIT"
 license-files = []
 
 [[licenses.clarify]]
-crate = "mollusk-test-utils"
-expression = "MIT"
-license-files = []
-
-[[licenses.clarify]]
-crate = "mollusk-harness"
+crate = "solana-axelar-mollusk-harness"
 expression = "MIT"
 license-files = []
