--- helpers/mollusk-test-utils/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "mollusk-test-utils"
-version = "0.1.0"
-edition = "2021"
-description = "Testing utilities for mollusk-based Solana program tests"
-license.workspace = true
-publish = false
-
-[lints]
-workspace = true
-
-[lib]
-crate-type = ["cdylib", "lib"]
-
-[dependencies]
-anchor-lang.workspace = true
-solana-sdk.workspace = true
-solana-sdk-ids.workspace = true
-mollusk-svm.workspace = true
-bincode.workspace = true
