+++ helpers/solana-axelar-mollusk-harness/Cargo.toml
@@ -1,8 +1,13 @@
 [package]
-name = "mollusk-harness"
+name = "solana-axelar-mollusk-harness"
 version = "0.1.0"
 edition = "2021"
-publish = false
+description = "Mollusk test harnesses for Axelar Solana programs"
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+rust-version.workspace = true
+readme = "README.md"
 
 [lints]
 workspace = true
@@ -19,13 +24,13 @@ mpl-token-metadata.workspace = true
 
 rand.workspace = true
 hex.workspace = true
+bincode.workspace = true
 libsecp256k1 = { workspace = true, features = ["hmac", "static-context"] }
 
 mollusk-svm.workspace = true
 mollusk-svm-programs-token.workspace = true
 spl-token.workspace = true
 
-mollusk-test-utils.workspace = true
 solana-axelar-std.workspace = true
 
 solana-axelar-gas-service = { workspace = true, features = ["cpi", "no-entrypoint", "devnet-amplifier"] }
