+++ programs/solana-axelar-gas-service/Cargo.toml
@@ -17,7 +17,7 @@ bytemuck.workspace = true
 [dev-dependencies]
 solana-sdk-ids.workspace = true
 mollusk-svm.workspace = true
-mollusk-harness.workspace = true
+solana-axelar-mollusk-harness.workspace = true
 solana-sdk.workspace = true
 borsh.workspace = true
