+++ programs/solana-axelar-gas-service/tests/add_gas.rs
@@ -1,8 +1,8 @@
 #![cfg(test)]
 
 use anchor_lang::prelude::ProgramError;
-use mollusk_harness::{GasServiceSetup, GasServiceTestHarness, TestHarness};
 use mollusk_svm::result::Check;
+use solana_axelar_mollusk_harness::{GasServiceSetup, GasServiceTestHarness, TestHarness};
 use solana_sdk::pubkey::Pubkey;
 
 #[test]
