+++ programs/solana-axelar-gas-service/tests/collect_fees.rs
@@ -2,8 +2,10 @@
 
 use anchor_lang::prelude::ProgramError;
 use anchor_lang::{InstructionData, ToAccountMetas};
-use mollusk_harness::{GasServiceSetup, GasServiceTestHarness, OperatorsSetup, TestHarness};
 use mollusk_svm::result::Check;
+use solana_axelar_mollusk_harness::{
+    GasServiceSetup, GasServiceTestHarness, OperatorsSetup, TestHarness,
+};
 use solana_sdk::pubkey::Pubkey;
 
 #[test]
