+++ programs/solana-axelar-gateway/tests/test_gateway_validate_message.rs
@@ -5,11 +5,11 @@ mod helpers;
 use helpers::*;
 
 use anchor_lang::{InstructionData, ToAccountMetas};
-use mollusk_harness::{GatewayTestHarness, TestHarness};
 use mollusk_svm::result::Check;
 use solana_axelar_gateway::{
     GatewayConfig, GatewayError, IncomingMessage, MessageStatus, ValidateMessageSigner,
 };
+use solana_axelar_mollusk_harness::{GatewayTestHarness, TestHarness};
 use solana_axelar_std::{Message, PayloadType};
 use solana_sdk::pubkey::Pubkey;
 
@@ -64,7 +64,9 @@ fn build_validate_message_ix(
     .expect("valid signing PDA");
 
     let (event_authority, _, _) =
-        mollusk_test_utils::get_event_authority_and_program_accounts(&solana_axelar_gateway::ID);
+        solana_axelar_mollusk_harness::get_event_authority_and_program_accounts(
+            &solana_axelar_gateway::ID,
+        );
 
     solana_sdk::instruction::Instruction {
         program_id: solana_axelar_gateway::ID,
@@ -167,7 +169,9 @@ fn validate_message_wrong_hash() {
     .expect("valid signing PDA");
 
     let (event_authority, _, _) =
-        mollusk_test_utils::get_event_authority_and_program_accounts(&solana_axelar_gateway::ID);
+        solana_axelar_mollusk_harness::get_event_authority_and_program_accounts(
+            &solana_axelar_gateway::ID,
+        );
 
     let ix = solana_sdk::instruction::Instruction {
         program_id: solana_axelar_gateway::ID,
@@ -212,7 +216,9 @@ fn validate_message_wrong_caller_pda() {
     .expect("valid signing PDA");
 
     let (event_authority, _, _) =
-        mollusk_test_utils::get_event_authority_and_program_accounts(&solana_axelar_gateway::ID);
+        solana_axelar_mollusk_harness::get_event_authority_and_program_accounts(
+            &solana_axelar_gateway::ID,
+        );
 
     let ix = solana_sdk::instruction::Instruction {
         program_id: solana_axelar_gateway::ID,
