+++ programs/solana-axelar-its/tests/test_its_outbound_transfer.rs
@@ -2,8 +2,8 @@
 #![allow(clippy::indexing_slicing)]
 
 use anchor_spl::token_2022;
-use mollusk_harness::{ItsTestHarness, TestHarness};
 use mollusk_svm::result::Check;
+use solana_axelar_mollusk_harness::{ItsTestHarness, TestHarness};
 
 #[test]
 fn user_interchain_transfer() {
