+++ programs/solana-axelar-memo/Cargo.toml
@@ -17,7 +17,7 @@ solana-axelar-its = { workspace = true, features = ["cpi"] }
 
 [dev-dependencies]
 mollusk-svm.workspace = true
-mollusk-harness.workspace = true
+solana-axelar-mollusk-harness.workspace = true
 solana-sdk.workspace = true
 solana-sdk-ids.workspace = true
 solana-axelar-std.workspace = true
